# gotham_like/docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Tessera Investigative Workbench",
  "description": "Local-first investigative intelligence & decision-support platform. Outputs are decision support: system inferences and analytical signals are never presented as established facts.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/provenance/{obj_id}": {
   "get": {
    "tags": [
     "governance"
    ],
    "summary": "Get Provenance",
    "operationId": "get_provenance_api_provenance__obj_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "obj_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Obj Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Provenance Api Provenance  Obj Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/provenance/{obj_id}/lineage": {
   "get": {
    "tags": [
     "governance"
    ],
    "summary": "Get Lineage",
    "operationId": "get_lineage_api_provenance__obj_id__lineage_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "obj_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Obj Id"
      }
     },
     {
      "name": "kind",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "pattern": "^(entity|relationship|event|source_record|signal)$",
       "title": "Kind"
      }
     },
     {
      "name": "direction",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "up",
        "down"
       ],
       "type": "string",
       "default": "up",
       "title": "Direction"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Lineage Api Provenance  Obj Id  Lineage Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/resolution/candidates": {
   "get": {
    "tags": [
     "governance"
    ],
    "summary": "List Candidates",
    "operationId": "list_candidates_api_resolution_candidates_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "decision",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "MATCH",
          "POSSIBLE_MATCH"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Decision"
      }
     },
     {
      "name": "review_status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Review Status"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "default": 100,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Candidates Api Resolution Candidates Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/resolution/compare": {
   "post": {
    "tags": [
     "governance"
    ],
    "summary": "Compare Entities",
    "operationId": "compare_entities_api_resolution_compare_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CompareBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Compare Entities Api Resolution Compare Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/resolution/candidates/{cand_id}/review": {
   "post": {
    "tags": [
     "governance"
    ],
    "summary": "Review Candidate",
    "operationId": "review_candidate_api_resolution_candidates__cand_id__review_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cand_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Cand Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ReviewBody"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Review Candidate Api Resolution Candidates  Cand Id  Review Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/rules": {
   "get": {
    "tags": [
     "governance"
    ],
    "summary": "List Rules",
    "operationId": "list_rules_api_rules_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "additionalProperties": true,
          "type": "object"
         },
         "type": "array",
         "title": "Response List Rules Api Rules Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/rules/{rule_id}": {
   "put": {
    "tags": [
     "governance"
    ],
    "summary": "Upsert Rule",
    "operationId": "upsert_rule_api_rules__rule_id__put",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "rule_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Rule Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RuleBody"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Upsert Rule Api Rules  Rule Id  Put"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/rules/{rule_id}/evaluate": {
   "post": {
    "tags": [
     "governance"
    ],
    "summary": "Evaluate Rule",
    "operationId": "evaluate_rule_api_rules__rule_id__evaluate_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "rule_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Rule Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Evaluate Rule Api Rules  Rule Id  Evaluate Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/signals": {
   "get": {
    "tags": [
     "governance"
    ],
    "summary": "List Signals",
    "operationId": "list_signals_api_signals_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "rule_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Rule Id"
      }
     },
     {
      "name": "entity_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Entity Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "default": 100,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Signals Api Signals Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/signals/{signal_id}": {
   "get": {
    "tags": [
     "governance"
    ],
    "summary": "Get Signal",
    "operationId": "get_signal_api_signals__signal_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "signal_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Signal Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Signal Api Signals  Signal Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CompareBody": {
    "properties": {
     "entity_a": {
      "type": "string",
      "title": "Entity A"
     },
     "entity_b": {
      "type": "string",
      "title": "Entity B"
     }
    },
    "type": "object",
    "required": [
     "entity_a",
     "entity_b"
    ],
    "title": "CompareBody"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ReviewBody": {
    "properties": {
     "decision": {
      "type": "string",
      "enum": [
       "CONFIRM",
       "REJECT"
      ],
      "title": "Decision"
     },
     "note": {
      "type": "string",
      "maxLength": 2000,
      "title": "Note",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "decision"
    ],
    "title": "ReviewBody"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}